fetched a new autoconf/automake environment from the KDE cvs server
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@203 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/newsignal/NewSignalPlugin.cpp
@@ -57,7 +57,7 @@ int NewSignalPlugin::interpreteParameters(QStringList &params)
     if (!ok) return -EINVAL;
 
     param = params[1];
-    m_rate = param.toDouble(&ok);
+    m_rate = (unsigned int)param.toDouble(&ok);
     ASSERT(ok);
     if (!ok) return -EINVAL;
